Moved javascript modules to modules folder
