+++ src/modules/class.js
