--- src/record.js
@@ -1,8 +1,0 @@
-import { TaskList } from './class.js';
-
-export const list = new TaskList();
-
-export const record = () => {
-  const taskname = document.querySelector('#addTask').value;
-  list.add(taskname);
-};
